+++ crates/e2e/src/lib.rs
@@ -53,11 +53,11 @@ pub use contract_results::{
     UploadResult,
 };
 pub use ink_e2e_macro::test;
+pub use ink_revive_types::evm::CallTrace;
 pub use node_proc::{
     TestNodeProcess,
     TestNodeProcessBuilder,
 };
-pub use pallet_revive::evm::CallTrace;
 #[cfg(feature = "sandbox")]
 pub use sandbox_client::{
     Client as SandboxClient,