+++ crates/e2e/src/subxt_client.rs
@@ -73,8 +73,8 @@ use ink_primitives::{
     DepositLimit,
     abi::AbiEncodeWith,
 };
+use ink_revive_types::evm::CallTrace;
 use jsonrpsee::core::async_trait;
-use pallet_revive::evm::CallTrace;
 use scale::{
     Decode,
     Encode,