+++ integration-tests/solidity-abi/sol-encoding/e2e_tests.rs
@@ -6,13 +6,13 @@ use ink::{
     primitives::DepositLimit,
 };
 use ink_e2e::ContractsRegistry;
+use ink_revive_types::ExecReturnValue;
 use ink_sandbox::{
     DefaultSandbox,
     Sandbox,
     api::prelude::*,
     frame_system::pallet_prelude::OriginFor,
 };
-use pallet_revive::ExecReturnValue;
 
 const STORAGE_DEPOSIT_LIMIT: DepositLimit<u128> = DepositLimit::UnsafeOnlyForDryRun;
 
@@ -114,7 +114,7 @@ impl ContractSandbox {
         data: Vec<u8>,
         origin: OriginFor<<DefaultSandbox as Sandbox>::Runtime>,
     ) -> ExecReturnValue {
-        <DefaultSandbox as ContractAPI>::call_contract(
+        let call_raw = <DefaultSandbox as ContractAPI>::call_contract(
             &mut self.sandbox,
             self.contract_addr,
             0,
@@ -124,7 +124,11 @@ impl ContractSandbox {
             STORAGE_DEPOSIT_LIMIT,
         )
         .result
-        .expect("sandbox call contract failed")
+        .expect("sandbox call contract failed");
+        ExecReturnValue {
+            flags: call_raw.flags,
+            data: call_raw.data,
+        }
     }
 }